refactor: move domains list to seperate file
