+++ ui/src/pages/organizations/details/security/index.tsx
@@ -8,61 +8,14 @@ import {
 } from "@raystack/apsara/v1";
 
 import styles from "./security.module.css";
-import { CheckCircledIcon, PlusIcon, TrashIcon } from "@radix-ui/react-icons";
-import { useEffect, useState } from "react";
+import { PlusIcon } from "@radix-ui/react-icons";
+import { useCallback, useEffect, useState } from "react";
 import { V1Beta1Domain } from "~/api/frontier";
 import { api } from "~/api";
 import { useOutletContext } from "react-router-dom";
 import { OutletContext } from "../types";
-import Skeleton from "react-loading-skeleton";
 import { BlockOrganizationSection } from "./block-organization";
-
-interface DomainItemProps {
-  domain: V1Beta1Domain;
-}
-
-const DomainItem = ({ domain }: DomainItemProps) => {
-  return (
-    <Flex className={styles["domains-list-item"]} justify="between">
-      <Flex gap={3}>
-        <Text size={3}>{domain?.name}</Text>
-        {domain.state === "verified" ? (
-          <CheckCircledIcon
-            color={"var(--rs-color-foreground-success-primary)"}
-          />
-        ) : null}
-      </Flex>
-      <IconButton size={3} data-test-id="delete-domain-button">
-        <TrashIcon />
-      </IconButton>
-    </Flex>
-  );
-};
-
-interface DomainListProps {
-  isLoading: boolean;
-  domains: V1Beta1Domain[];
-}
-
-const DomainList = ({ isLoading, domains = [] }: DomainListProps) => {
-  return isLoading ? (
-    <Flex direction="column" className={styles["domains-list"]}>
-      {[...Array(3)].map((_, index) => (
-        <Skeleton
-          key={index}
-          height={20}
-          style={{ margin: "var(--rs-space-5) 0" }}
-        />
-      ))}
-    </Flex>
-  ) : domains.length ? (
-    <Flex direction="column" className={styles["domains-list"]}>
-      {domains.map((domain) => (
-        <DomainItem key={domain?.id} domain={domain} />
-      ))}
-    </Flex>
-  ) : null;
-};
+import { DomainsList } from "./domains-list";
 
 const AddDomainSection = () => {
   return (
@@ -95,10 +48,12 @@ export const OrganizationSecurity = () => {
   const [domains, setDomains] = useState<V1Beta1Domain[]>([]);
   const [isDomainLoading, setIsDomainLoading] = useState(false);
 
-  async function fetchDomains(orgId: string) {
+  const fetchDomains = useCallback(async () => {
+    if (!organizationId) return;
     try {
       setIsDomainLoading(true);
-      const response = await api?.frontierServiceListOrganizationDomains(orgId);
+      const response =
+        await api?.frontierServiceListOrganizationDomains(organizationId);
       const data = response?.data?.domains || [];
       setDomains(data);
     } catch (error) {
@@ -106,17 +61,21 @@ export const OrganizationSecurity = () => {
     } finally {
       setIsDomainLoading(false);
     }
-  }
+  }, [organizationId]);
 
   useEffect(() => {
-    if (organizationId) fetchDomains(organizationId);
-  }, [organizationId]);
+    fetchDomains();
+  }, [fetchDomains]);
 
   return (
     <Flex justify="center" className={styles["container"]}>
       <Flex className={styles["content"]} direction="column" gap={9}>
         <AddDomainSection />
-        <DomainList isLoading={isDomainLoading} domains={domains} />
+        <DomainsList
+          isLoading={isDomainLoading}
+          domains={domains}
+          fetchDomains={fetchDomains}
+        />
         <Separator />
         <BlockOrganizationSection />
       </Flex>
